refactor: slice large worker internals
